Rename el userdata provider to centos
