+++ pkg/cloudprovider/provider/hetzner/provider.go
@@ -51,7 +51,7 @@ func getNameForOS(os providerconfig.OperatingSystem) (string, error) {
 	switch os {
 	case providerconfig.OperatingSystemUbuntu:
 		return "ubuntu-16.04", nil
-	case providerconfig.OperationSystemEnterpriseLinux:
+	case providerconfig.OperatingSystemCentOS:
 		return "centos-7", nil
 	}
 	return "", providerconfig.ErrOSNotSupported
