+++ pkg/providerconfig/types.go
@@ -20,9 +20,9 @@ var (
 type OperatingSystem string
 
 const (
-	OperatingSystemCoreos          OperatingSystem = "coreos"
-	OperatingSystemUbuntu          OperatingSystem = "ubuntu"
-	OperationSystemEnterpriseLinux OperatingSystem = "el"
+	OperatingSystemCoreos OperatingSystem = "coreos"
+	OperatingSystemUbuntu OperatingSystem = "ubuntu"
+	OperatingSystemCentOS OperatingSystem = "centos"
 )
 
 type CloudProvider string
