+++ pkg/userdata/centos/userdata_test.go
@@ -1,4 +1,4 @@
-package el
+package centos
 
 import (
 	"net"
