+++ pkg/userdata/userdata.go
@@ -6,9 +6,9 @@ import (
 
 	machinesv1alpha1 "github.com/kubermatic/machine-controller/pkg/machines/v1alpha1"
 	"github.com/kubermatic/machine-controller/pkg/providerconfig"
+	"github.com/kubermatic/machine-controller/pkg/userdata/centos"
 	"github.com/kubermatic/machine-controller/pkg/userdata/cloud"
 	"github.com/kubermatic/machine-controller/pkg/userdata/coreos"
-	"github.com/kubermatic/machine-controller/pkg/userdata/el"
 	"github.com/kubermatic/machine-controller/pkg/userdata/ubuntu"
 )
 
@@ -16,9 +16,9 @@ var (
 	ErrProviderNotFound = errors.New("no user data provider for the given os found")
 
 	providers = map[providerconfig.OperatingSystem]Provider{
-		providerconfig.OperatingSystemCoreos:          coreos.Provider{},
-		providerconfig.OperatingSystemUbuntu:          ubuntu.Provider{},
-		providerconfig.OperationSystemEnterpriseLinux: el.Provider{},
+		providerconfig.OperatingSystemCoreos: coreos.Provider{},
+		providerconfig.OperatingSystemUbuntu: ubuntu.Provider{},
+		providerconfig.OperatingSystemCentOS: centos.Provider{},
 	}
 )
